Create library, add books and book grid display
Library is working as intended, with book adding capabilities via
HTML forms accesible with the "Add book" button, as well as displaying
added books in a responsive grid. The project is developed for Desktop
use, as the course hasn't yet covered responsive design with CSS for
Mobile, but will return to it to make it mobile-friendly in the future.